requirements.txt
# Updated for latest Django and MongoEngine
Django>=4.2,<5.0
mongoengine>=0.27.0
# djongo removed - using mongoengine directly
Pillow
amqp
# anyjson removed - incompatible with Python 3.8+, use simplejson instead
billiard
biplist
celery>=5.3.0
defusedxml
django-debug-toolbar>=4.2.0
django-tastypie>=0.14.5
django-tastypie-mongoengine
django-mongoengine
future
kombu
lxml
pymongo>=4.5.0
pyparsing
python-dateutil
python-ldap
python-magic>=0.4.6
python-mimeparse
pytz
pyyaml
requests
simplejson
six
# ushlex - incompatible with Python 3, use shlex from stdlib instead
olefile
pycryptodome
# pyimpfuzzy - requires fuzzy.h, skipping for now
# pydeep>=0.4 - requires fuzzy.h, skipping for now
chardet
qrcode
